Fix ABI test pipeline
The container image is changed from Rocky 8 to CentOS 7 to install zstd-static.
The repository to compate with is changed to Arenadata fork.
Git is installed from the endpointdev repository to clone with submodules.
Lynx is now installed from the main repository, because it is located in the
CentOS 7 main repository and there is no powertools repository in CentOS 7.