Fix deadlock if the cross slice SharedScan consumer is on slice 0 (#623)
When the consumer of the cross slice SharedScan is in slice 0 and the producer
is in another slice, a deadlock may occur. The deadlock could happen because QD
does not complete slice 0 until other slices will be completed. But the
producer's slice cannot finish until it receives notification from the consumer
about the completion of reading. The consumer does not send this notification
before all QEs are completed.

This patch fixes the deadlock by creating a list of cross-slice consumers for
the current slice and sending notifications before waiting for QEs completion.
We can omit the share_lk_ctxt check for NULL since only cross slice Shared Scans
are added to the list.