Мake the gpexpand test more stable.
In some gpexpand tests, checks if the expand is interrupted after a set time,
but sometimes gpexpand may expand faster than the set time and the test will
fail.

This patch adds a 5-second sleep after the expansion of the first table to
stimulate a long expansion without increasing the amount of data. To add sleep,
a fault injector in the finish_heap_swap is used, which occurs during the
expansion of the table.