Fix the check for address availability when setting up UDP sockets in IPv6 disabled environments (#679)
This patch brings changes introduced in 7f3c91f (which backported 70306db,
1ee34a4) and fixes them by correctly handling IPv6 addresses.

In these patches, a fix was created for UDPIFC that reworked
SendDummyPacket() by removing the unnecessary call to getaddrinfo() to avoid
using malloc() in a signal handler, making setupUDPListeningSocket cache the
address of the socket used for the UDPIFC interconnect when it was first
created.

In IPv6 disabled environments, bind() with address of an IPv6 wildcard will
result in binding to both IPv6 and IPv4 (since IPv6 is disabled, only IPv4 will
be bound) on all interfaces, and it won’t fail even if IPv6 is disabled. This
mistakenly passed the check and cached address with IPv6 address family. In IPv6
disabled environments, any connect() or sendto() that reuse this address
would fail. This fail occured when we wanted to send a dummy packet with
SendDummyPacket() using the cached address.

Add AI_ADDRCONFIG flag to getaddrinfo() hints to force it to check for
availability of the address families. This flag will prevent getaddrinfo()
from returning addresses that belong to the IPv6 family when IPv6 is not
available on the host machine.

This problem is present only when INTERCONNECT_ADDRESS_TYPE_WILDCARD is used
as the interconnect address type. It does not affect 7.X, because the default
setting for interconnect address type was changed to
INTERCONNECT_ADDRESS_TYPE_UNICAST.

This is a rework of cherry-picked commit 7f3c91f.

Co-authored-by: Marbin Tan tmarbin@vmware.com