add correct baseline ignore file for ABI tests
The 6.26.2 sync contains changes (b72ff6b) that break ABI. Later in
af90388 ignore file was added for version 6.26.1. But this tag is absent
in our fork and moreover, we should validate ABI changes against our
latest tag - 6.26.0_arenadata54. Because otherwise we got all changes
between upstream tag and this PR that includes our changes. What is why
I've duplicated this changes to the ignore file for our tag.