Optimize the Explicit Redistribute application logic (#666)
Previously, if a plan contained any motions and a ModifyTable node, Explicit
Redistribute Motion would be added under the ModifyTable to send tuples back to
a specific segment based on gp_segment_id.

We only really need an Explicit Redistribute Motion if the tuples that we are
going to modify were sent to a different segment by a Motion, because DELETE and
UPDATE use ctid to find tuples that need to be modified, and ctid only makes
sense in its original segment.

This patch updates the logic: add Explicit Redistribute Motion only if there is
a motion that distributes the relation we are going to update and this motion is
between the scan and the ModifyTable.

UPDATE/DELETE operations are represented by ModifyTable node. We expand
ApplyMotionState with a list of resulting relations retrieved from the
ModifyTable node, and a motion counter. A counter is used to account for the
possibility of multiple motions before the scan.

When we encounter a scan node, we search for it's scanrelid (the index of
relation) in our list of resulting relations. If the list contains this
scanrelid, we have found the scan on the table that's going to be modified by
ModifyTable, and remember the motion count. We ignore scans under InitPlans,
since there will not be any scan nodes that perform a scan on the same range
table entry as ModifyTable under InitPlans.

When we encounter the Explicit Redistribute Motion itself, we add it based on
the counter. If the motion counter is not 0, set the we can't elide the Explicit
Redistribute Motion, since the tuples that are going to be modified were
previously sent to a different segment by a motion. Otherwise, if motion counter
is 0, we can elide it.

The solution is based on a46400e.