Fix resource group tests (#883)
Patch 52bbf95d91155517a711b193bbffbaec880e7389 increased memory consumption on
the coordinator and caused the resgroup_cpu_rate_limit test to fail.

This patch adds the ability to create a demo cluster on multiple hosts, which
allows us to create an environment close to the upstream.
Also disables the ORCA planner for resource group tests for two reasons. First,
starting with commit 39de39eb6edb0d0f94f533e96e04b66d6c790d14, upstream runs
tests without it. Secondly, in the resgroup_cpu_rate_limit test, there is a
problem that when the CPU is limited, ORCA can plan a query for a long time,
which leads to the test failing.