Make frozen_insert_crash test stable (#893)
Commit 1502bef introduced a new flaking test. The test expects a panic to occur
in the process that performs the insert. But sometimes the panic happens first
in process which started by dtx recovery process that periodically checks for
orphaned transactions. Which leads to an error in the test. The solution is to
panic only the required process, which is done by extracting session from
pg_stat_activity by query.