Fix SIGSEGV while executing lateral subqueries with aggregate functions (#886)
Problem description:
Some queries with a lateral join cause a segmentation fault. It happens if the
query has a subquery with aggregate functions and without a GROUP BY clause or
it has more than one level nesting.

Cause:
Postgres planner for these queries creates a plan with slices, and the lateral
variable is accessed from a different slice. Lateral variables can't be sent
across motions. So, it causes a segmentation fault.

Fix:
Fix is to force postgres planner to bring such queries to a single query
executor by setting 'force_singleQE' in the PlannerConfig structure. Previously,
this flag was set only for queries with lateral variables and a subquery with
GROUP BY / LIMIT / DISTINCT clauses. Now this condition is complemented by
checking the presence of:
 - aggregate functions;
 - set operations that can cause motion (INTERSECT, EXCEPT, UNION without ALL);
 - recursive CTEs that also can cause motion.
Plus recursive check for all CTEs and SUBQUERY RTEs was added.