Add support for table aliases in ORCA (#780)
Currently ORCA generated query plan does not show aliases, unlike original
PostgreSQL planner. Usually it is much more descriptive to have the alias in the
plan, which are specified by user in the query and might have meaning.

This patch adds table aliases support to ORCA. Table aliases are added to both
internal structures and DXL schema and for debug representation of internal
objects. Since explain command uses RangeTableEntry fields of PlannedStmt to
extract the alias, we also fill rte->alias field during PlannedStmt creation.

Table aliases are widely used in the tests, so simple adding them requires lots
of tests changes and patch becomes rather huge. To eliminate this, GUC parameter
optimizer_enable_table_alias has been introduced. Default value is on, it
enables table aliases. Setting it to off makes aliases not used, as before the
patch.

The optimizer_enable_table_alias parameter intended mainly for automated tests,
so tests have to set it explicitly for all the cases. This is done in two steps:
1. When optimizer=on is used, i.e. we are running ORCA-enabled tests,
the OPTIMIZER_ENABLE_TABLE_ALIAS environment variable is set to off. 
2. At the demo cluster startup, in demo_cluster.sh, if
OPTIMIZER_ENABLE_TABLE_ALIAS is set, the line optimizer_enable_table_alias=off
is appended to the file clusterConfigPostgresAddonsFile which is used to
distribute configuration across demo cluster. Basically,
optimizer_enable_table_alias is needed only for the master node and could be set
on per-connection basis, but this approach is simple, consistent and has minimum
required changes.