Eliminate unnecessary variables in the DecodeDateTime() function
The *tzp variable was not utilized anywhere in the DecodeDateTime() code, so
these assignments were redundant. Some callers of DecodeTimezone() depended on
the *tzp being non-NULL within that function.

Extend DecodeTimezone() to accept a NULL value as a second parameter. Use that
NULL value in DecodeDateTime() to indicate that it should not be modified.
Eliminate the need for t and *tzp in DecodeDateTime() and simplify the code.