pg_upgrade: check for views using functions with changed signatures
Views that use functions with changed signatures will cause
upgrade to fail. This happens during metadata restore on the
target cluster because pg_restore will error trying to create a view
using functions with changed arguments and/or return types.
This is not ideal as we could be many hours into upgrade before
a view using a function with changed signature causes pg_upgrade to fail
This check calls a support function on the source cluster to check if
views using such functions with changed signature exist.

(cherry picked from commit 11973217973ec7decfdf6d411fce70cedb160135)

Changes compared to the original commit:
- Use fully qualified names
  and remove queries that change the schema.
- Properly format error message.
- Close created file.
- Unhardcode FirstNormalObjectId.