Replace memcpy with memmove in SWAPN macro (#918)
memcpy was used to copy overlapping buffers in the SWAPN macro. For memcpy,
according to its specification, "if the objects overlap, the behavior is
undefined". Thus, it may potentially cause errors. The memcpy function is
replaced with memmove, which also copies data from one buffer to another one,
but the buffers may overlap according to its specification.