Do not treat interruption of LDAP requests by signals as an auth error. (#899)
When authorizing via LDAP, the ldap_simple_bind_s function may return an error.
This occurs when an internal call to the ldap_int_poll utility function is
interrupted by a signal. The openldap library can handle this properly using the
LDAP_OPT_RESTART option. Setting this option to LDAP_OPT_ON causes the select()
system call to be restarted when interrupted by a signal. The default value is
LDAP_OPT_OFF. This patch sets the value to LDAP_OPT_ON (also it requires
explicitly specifying the -llber library to build) after ldap is initialized.

Tests are not provided because they require a separate LDAP container and the
error is not reproducible in a stable manner.