Fix segfault in get_ao_compression_ratio (#907)
Problem:
Segfault happened in the queries with 'get_ao_compression_ratio' function in the
target list and the argument for this function that was taken from a table. The
problem was reproduced in the release build. In the development build the same
queries led to assertion.

Cause:
Function 'get_ao_compression_ratio' executes a query on
'pg_aoseg.pg_aoseg_' table to calculate the compression ratio. A function
that accesses hash-distributed or randomly distributed tables must be executed
on coordinator. But the planner created plans with the
'get_ao_compression_ratio' function which was executed on segments. Inside
'get_ao_compression_ratio' there was an assert to check this case, but on the
release build the assert was removed. So, the code of the function was executed
on a segment and got unexpected results from 'pg_aoseg.pg_aoseg_', that led to a
segfault.

Fix:
Assert is changed to a runtime check. Now, if the planner creates a plan with
the function which is executed on segments, the user will get a proper error
message. It is also done for the similar functions 'gp_update_ao_master_stats'
and 'get_ao_distribution'. Plus, unreachable code is removed from internal
functions.

Note: it seems that a fully proper way would be to mark these functions as
PROEXECLOCATION_MASTER instead of PROEXECLOCATION_ANY, but that would require
changing of the system catalog, and it is acceptable only between major versions
update. So this approach is not used.

Note2: for an empty column-oriented table 'get_ao_compression_ratio' returns -1,
but for an empty row-oriented table 'get_ao_compression_ratio' returns 1. This
behavior is inconsistent, but it is not altered by this commit as not related
to the original issue. It has a reflection in the added test case. Be aware.

(cherry picked from commit 3dea58de92a9b040fe5ce194048a97f55bbdb71b)