Fix segfault when copying dropped column ACLs to a new partition
While copying columns ACLs when adding a new partition, we assumed that the
attribute numbers in the source and target relation would match, which is not
true in the case of the dropped columns in the original relation. In this case,
we would try to get an ACL for a non-existent column in the target table, which
would lead to a segfault.

This patch adds a separate iterator for the list of columns in the target
relation to obtain the correct skipping of dropped columns in the original
relation and obtain the correct attribute number for the target relation. We
assume that there are no dropped columns in the target relation, since this is a
newly created relation.

(cherry picked from commit 13482adee24069532915ce849ab2dc0019cf15e5)

Changes comparing to the original commit:

1. Additional cleanup is added before the test start, as the role used in the
test couldn't be dropped (and re-created later), because the same role was used
in the preceding test and there were some objects depending on this role. So we
remove these objects now before starting the test.

2. Now the partition table may have different number of system columns comparing
to the top level parent partitioned table. The original code didn't handle this
case, and it resulted in an assert when creating a partitioned table. The code
to handle different number of system columns is added.

3. Tests are updated to cover cases with different number of system columns

Co-authored-by: Roman Eskin <r.eskin@arenadata.io>