Implement default transformation flags for gpfdist (#934)
Currently, gpfdist transformation can only be applied to table source on
external table level. It can be done by specifying #transform tag. Sometimes it
may be not handy to specify the same transformation for all source files. The
goal of the patch is to implement new flags with the default transformation name
for input and output transformations. The default transformation(s) can be set
at gpfdist start and applied only if table level transformation is not
specified.

New flags are a subject of only basic check. The idea was to keep the same
behavior and check transformation existence only at runtime.
The test shows a behavior described in patch goal.