Reimplement address availability check for UDP interconnect
This patch reimplements changes introduced in 5beccaa (which backported 70306db,
1ee34a4) and fixes them by correctly handling IPv6 addresses, including
link-local ones.

In these patches, a fix was created for UDPIFC that reworked SendDummyPacket()
by removing the unnecessary call to getaddrinfo() to avoid using malloc() in a
signal handler, making setupUDPListeningSocket cache the address of the socket
used for the UDPIFC interconnect when it was first created.

The original patch (5beccaa) didn’t account that AI_ADDRCONFIG flag does not
protect us from link-local addresses (and other non-global addresses), which are
issued automatically upon connecting to some networks. These addresses cause a
false positive even with that flag.

When IPv6 is disabled, all IPv6 addresses are removed from the network
interfaces. However, upon reconnecting to the network, these addresses reappear,
even though IPv6 remains disabled, with a single exception being loopback
interface, IPv6 addresses of which do not reappear.

AI_ADDRCONFIG usage is replaced with a manual check, which looks for the first
available IPv6 address and simultaneously makes sure that the loopback interface
has an IPv6 address assigned.

This is a rework of commit 5beccaa.