Fix invalid attribute number when executing shared CTE (#939)
Previously, the planner encountered a problem with the handling of the target
list for the ShareInputScan. Typically, the target list of ShareInputScan is
directly copied from it's outer plan during the optimization phase, since it
wasn't evalutated and was needed only for EXPLAIN. However, when dealing with a
DML statement that includes a RETURNING clause, the behavior was incorrect, due
to the ModifyTable's targetlist only being initialized during the execution
phase. This led to an error due to a difference between the expected and actual
number of columns when computing the result tuple's descriptor from the
targetlist, as it was incorrectly copied as NIL.

This patch addresses this issue by setting up a dummy targetlist for ModifyTable
while in optimization phase to ensure that the ShareInputScan's targetlist is
correctly copied later. Since the ModifyTable's targetlist is now initialized,
an existing workaround used for Motions is removed.