ORCA does not fully support using CTE with replicated tables  (#942)
Add an extra test case for planning queries with ORCA with replicated tables
using CTE.

There was a well-known problem in ORCA with planning distributed queries with
replicated tables using CTE. It was addressed several times in both adb and
upstream (PRs 13833 - cdd532f,
PR 14896, PR 13728, ADBDEV-2411), but only the upstream commit 24a54a7
which included in 6.27 provided a more complete solution.

This patch adds one more test case for the aforementioned problem, which was not
appeared in the tests but was solved. In the particular query in the test in
this patch ORCA could not build a plan due to too heavy restrictions applied to
CPhysicalSequence by using CDistributionSpecNonSingleton.