Update python to 3.11 version in Dockerfile
Upgrade the Python version in the Docker container to 3.11 since we need this
since 2765123a5c04590ea5e093c78959cd6e5abbf043.
Install psutil from pypi since there is no psutil package for Python 3.11 in the
repository.
Install the pip package, as it is not specified as a dependency for the
Python 3.11 base package.