Skip already freed tuples in tuplestore_end (#977)
If we have insufficient memory to execute a query, postgres spills some data to
disk. However, if we hit spill file size limit, a query will be cancelled. When
spilling tuples to disk, postgres clears them from memory. However, it doesn't
mark them as inaccessible. So, if an error appears, tuplestore_end cycles
through all tuples and tries to clear them. Attempting to clear an already
freed tuple tuplestore_end causes segfault. Segfault happens because gpdb has
an additional property in memory chunk header which is sharedHeader. If a tuple
was once freed then sharedHeader is set NULL in AllocSetFree. The next attempt
to free a tuple dereferences sharedHeader and causes segfault. Segfault causes
postgres to go into segfault handler and drop current stack. This happens in
the middle of working loop in dumptuple so the loop never ends. The next
attempt to clear tuplestore tries to free the same tuples once again which
causes segfault. This patch counts the number of freed tuples so that
tuplestore_end skips them in case of an error.

Ticket: ADBDEV-5729