COPY isn't interrupted when client gone (#989)
Add interrupt checks to COPY FROM command. For this fread was replaced with
WaitLatchOrSocket and read with a timeout. This way the interrupts are checked even
inside long-running COPY commands with very slow connections, and the COPY is
interrupted successfully.

Adding a test for this patch required a modification for isolation tests Python
script to support terminating a connection without waiting for the command to
finish.

Ticket: ADBDEV-5789