Fix tests on Ubuntu docker image (#1018)
Add allure-behave, fakeroot and sudo to run behave tests. Add krb5-kdc and
krb5-admin-server to run regression tests.
Add "init: true" to kill postgres zombie processes. The pgrep utility found
these processes and their existence was interpreted by behave tests as GPDB was
running.
Set the limit on the number of open files to 65535. Without this setting
the limit was 64000. It caused to warning in gpinitsystem, so behave tests did
not pass.
The entrypoint value is changed to "sleep infinity" for simplification and
unification with GPDB 7.