Fix GUC sync leaving broken temp tables (#1049)
Before this patch, if an error happened during GUC sync (for example a segment
was disconnected) the gang was destroyed but the session wasn't reset, so
temporary tables weren't cleaned up. This left a temporary table left on
coordinator, but not on segments, resulting in segment errors. This patch
fixes this by enabling session reset for errors during GUC syncs.

Ticket: ADBDEV-6260