Fix hanging of cluster stop in coordinator standalone mode with extension (#1052)
Problem:
In case cluster was started with only coordinator instance without segments and
FTS, and with enabled extension that dispatched a query, that involved segments,
attempt to stop the cluster failed with hanging.

Root cause:
When stopping a cluster, SIGTERM is sent to the postmaster. Postmaster sends in
turn SIGTERM to its child processes, including bg workers. Extension replaces
the standard handler for SIGTERM signal with its own, which just sets a flag,
that is supposed to handled by main loop of extension bg worker. But bg worker
process was busy with executing of a query, that couldn't be complete because it
was caught in an infinite loop in FtsNotifyProber(), that couldn't be broken as
there is no FTS process.

Fix:
The patch adds a check in the FtsNotifyProber() that FTS process is started. If
there is no FTS process, then just return from the FtsNotifyProber().

Tests are not presented, as the issue was reproduced only with a particular
extension.

The change is a part of GPDB 7x commit 35e95b9
The commit itself is too large to cherry-pick the entire commit.

Ticket: ADBDEV-6311

Co-authored-by: Huansong Fu <fhuansong@vmware.com>
Co-authored-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>
Co-authored-by: Jimmy Yih <jyih@vmware.com>