Add signal handlers to FtsProbe and WalReceiver (#1059)
Add signals handlers for SIGSEGV, SIGILL and SIGBUS to FtsProbe and
WalReceiver, similar to GPDB 6. The standard handler is used (the one that
logs a stacktrace) with corresponding process names.

Unlike GPDB 6, CdbProgramErrorHandler is not used for FtsProbe, because
FtsProbe does not use threads and therefore the standard handler is enough.

There are no tests added because testing for stacktraces appearing in logs is
complicated and unnecessary. To test this manually:

1. Run GPDB demo cluster with mirrors
2. Send a signal to FtsProbe or WalReceiver process with
pkill -SIGSEGV -f ftsprobe or pkill -SIGSEGV -f walreceiver.
3. Check GPDB logs for stacktraces

Ticket: ADBDEV-6348