Support creating external table with default clause (#983)
If the database contains an external table with a DEFAULT or NOT NULL
constraints (for example, created because of a serial type field), it is
possible for gpbackup to output a CREATE EXTERNAL TABLE statement with DEFAULT
and NOT NULL specified in it directly. Because DEFAULT and NOT NULL constraints
are already supported by all other parts of ADB, and because ALTER TABLE
already allows adding them to external tables, this patch allows the CREATE
EXTERNAL TABLE syntax to specify them too, to make that functionality directly
accessible. The change affects both readable and writable external tables.

NOT NULL constraints on readable external tables are ignored and issue a
warning because we cannot check the external table data, it might actually
contains nulls. On writable external tables NOT NULL constraints are enabled
and enforced.

SET NOT NULL and DROP NOT NULL is also allowed for ALTER EXTERNAL TABLE now,
the same way SET DEFAULT was already enabled, however they are ignored on
readable tables.

Ticket: ADBDEV-4724