Prohibit starting of postmaster with EXECUTE role. (#1065)
Prohibit starting of postmaster with EXECUTE role.

Normally segment node is started in utility role and execute role is used only
when backend process is started. However if one tries to start segment directly
in the execute mode using pg_ctl, this leads to SIGSEGV in the postgres process.

It happens because pointer to shared memory flag shmCleanupBackends is left
uninitialized. This pointer initialized in the function tmShmemInit() only for
GP_ROLE_DISPATCH and GP_ROLE_UTILITY and the flag itself is meaningful only for
GP_ROLE_DISPATCH.

This patch prohibits start of the postmaster in execute mode and raises error in
this case. To ensure that flag shmCleanupBackends is initialized correctly an
assert was added.

This scenario assumes unexpected startup procedure and quite specific use case
not falling into regular regression tests. Automated testing of this case would
require quite complicated TAP test and does not bring any value to normal use
cases. Considering this, it seems not feasible to make a specific automated test
for this scenario.

---------

Co-authored-by: Viktor Kurilko <v.kurilko@arenadata.io>