Fix assertion about exceeding maximum runs in mk-sorter (#1082)
When there is insufficient memory, the mk-sorter uses a larger number of
sorting runs. And at some point it may exceed the maximum allowed number of
runs, which is 16383. This limitation is caused by 14-bit storage. The
mke_set_run and mke_set_lv functions use an assert that checks for going beyond
the maximum value. But this assert is redundant. Firstly, the release build
simply throws an error in this case. And secondly, increasing memory solves the
problem. Also, disabling the GUC gp_enable_mk_sort solves the problem. So we
can simply always throw an error and the user can solve it by increasing memory
or using another sorter.

Ticket: ADBDEV-6484